Convert to new csproj format
