+++ Nfield.Quota.Tests/Assets/Asset.cs
