+++ Nfield.Quota/Persistence/QuotaFrameContractResolver.cs
