+++ Nfield.Quota/Persistence/QuotaFrameEncoder.cs
