+++ Nfield.Quota/PreOrderQuotaFrameTraverser.cs
