+++ Nfield.Quota/Properties/AssemblyInfo.cs
