+++ Nfield.Quota/QuotaFrameVariable.cs
